backend/requirements.txt
alembic==1.10.2
apispec==6.3.0
APScheduler==3.10.0
atlassian-python-api==3.35.0
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
Cerberus==1.3.4
certifi==2024.2.2
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.14
exceptiongroup==1.1.2
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.3.0
Flask-Login==0.6.2
flask-marshmallow==0.14.0
google-api-core==2.28.1
google-api-python-client==2.92.0
# Updated to support urllib3 2.x (required for security fixes)
google-auth==2.43.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.71.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
limits==3.5.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mdurl==0.1.2
mistune==3.0.1
oauthlib==3.2.2
ordered-set==4.1.0
packaging==23.1
pdpyras==4.5.2
pluggy==1.2.0
# Security fix: SNYK-PYTHON-PROTOBUF-10364902 (CVE-2025-4565) - Uncontrolled Recursion
# Requires >= 6.31.1 (or >= 4.25.8 or >= 5.29.5)
protobuf==6.33.0
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.1.0
pytest==7.2.1
pytest-env==0.8.1
pytest-mock==3.10.0
pytest-sqlalchemy-mock==0.1.4
python-dotenv==1.0.0
pytz==2023.2
PyYAML==6.0
referencing==0.29.1
# Updated to support urllib3 2.x (required for security fixes)
requests==2.32.5
requests-mock==1.10.0
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.8.8
rsa==4.9
six==1.16.0
slack-bolt==1.17.0
slack-sdk==3.20.2
SQLAlchemy==1.4.46
tomli==2.0.1
typing_extensions==4.7.1
tzlocal==5.0.1
uritemplate==4.1.1
# Security fix: SNYK-PYTHON-URLLIB3-14192442 (CVE-2025-66471) - Improper Handling of Highly Compressed Data
# Security fix: SNYK-PYTHON-URLLIB3-14192443 (CVE-2025-66418) - Allocation of Resources Without Limits or Throttling
# Requires >= 2.6.0
urllib3==2.6.0
waitress==3.0.2
# Security fix: SNYK-PYTHON-WERKZEUG-6808933 (CVE-2024-34069) - Remote Code Execution
# Requires >= 3.0.3
Werkzeug==3.1.3
wrapt==1.15.0
# Security fix: SNYK-PYTHON-SETUPTOOLS-7448482 (CVE-2024-6345) - Code Injection
# Requires >= 70.0.0 (pinned to exact version to prevent drift)
setuptools==80.9.0
